split up cell functions
